requirements.txt
numpy==2.1.3
requests==2.32.3
sympy==1.13.3
tqdm==4.67.1
pillow==11.0.0